Include 'Arbitrary Novella' in library. Benchmarks
Benchmarks require some kind of generator. For now, it seems it's easier
to put Gen instances in main library, and re-use them in both tests and
benchmarks.

Drawbacks:

- main library depends on QuickCheck,
- tests and benchmarks do not depend on the library, instead they have
  direct access to src...
- ...which leads to duplication in .cabal file: both tests and
  benchmarks must include appropriate `other-modules` including stuff
  from src.

The above drawbacks make this a regression in the code quality. The good
solution is to separate the Gen instances into separate package.